calculo de descontos funcional, e agora descontos devem ser um inteiro (tipo 10, 15 50 de desconto), ja que fica mais intuitivo do que ja usar o double de cara (0.10, 0.15, 0.5)

+++ src/main/java/com/ufcg/si1/model/Categoria.java
@@ -16,7 +16,7 @@ public class Categoria {
 	private String nome;
 
 	@Column
-	private double desconto;
+	private int desconto;
 
 
 	public Categoria() {
@@ -24,7 +24,7 @@ public class Categoria {
 	}
 
 
-	public Categoria(List<Produto> produtos, String nome, double desconto) {
+	public Categoria(List<Produto> produtos, String nome, int desconto) {
 		this.nome = nome;
 		this.desconto = desconto;
 	}
@@ -55,7 +55,7 @@ public class Categoria {
 	}
 
 
-	public void setDesconto(float desconto) {
+	public void setDesconto(int desconto) {
 		this.desconto = desconto;
 	}
 }
